Rename ExampleUI folder to example-ui. Use named exports
