+++ packages/nextjs/components/example-ui/ContractInteraction.tsx
@@ -5,7 +5,7 @@ import { ArrowSmallRightIcon, XMarkIcon } from "@heroicons/react/24/outline";
 import { useState } from "react";
 import { useScaffoldContractWrite } from "~~/hooks/scaffold-eth";
 
-export default function ContractInteraction() {
+export const ContractInteraction = () => {
   const [visible, setVisible] = useState(true);
   const [newGreeting, setNewGreeting] = useState("");
 
@@ -78,4 +78,4 @@ export default function ContractInteraction() {
       </div>
     </div>
   );
-}
+};
