+++ packages/nextjs/components/example-ui/assets/CopyIcon.tsx
