+++ packages/nextjs/components/example-ui/assets/DiamondIcon.tsx
