+++ packages/nextjs/components/example-ui/assets/HareIcon.tsx
